+++ sdk/azcore/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 * Added `ClientOptions.APIVersion` field, which overrides the default version a client
   requests of the service, if the client supports this (all ARM clients do).
+* Added package `tracing` that contains the building blocks for distributed tracing.
+* Added field `TracingProvider` to type `policy.ClientOptions` that will be used to set the per-client tracing implementation.
 
 ### Breaking Changes